+++ src/Stage.ts
@@ -737,7 +737,10 @@ export class Stage extends Container<Layer> {
 
     // always call preventDefault for desktop events because some browsers
     // try to drag and drop the canvas element
-    if (evt.cancelable && eventType !== "touch") {
+    // TODO: are we sure we need to prevent default at all?
+    // do not call this function on mobile because it prevent "click" event on all parent containers
+    // but apps may listen to it.
+    if (evt.cancelable && eventType !== 'touch') {
       evt.preventDefault();
     }
   }